fix(utils): use unitDisplay long in formatDuration for consistent output
Switch from 'short' to 'long' to avoid ICU inconsistencies across
environments ('mo.' vs 'mths'). Output is now stable: '1 year 9 months',
'1 ano 9 meses', etc.

+++ apps/site/src/utils/formatters/formatDuration.ts
@@ -15,7 +15,7 @@ export function formatDuration(
     new Intl.NumberFormat(locale, {
       style: 'unit',
       unit,
-      unitDisplay: 'short',
+      unitDisplay: 'long',
     }).format(value);
 
   const parts: string[] = [];
